+++ checkpoint/orbax/checkpoint/_src/path/gcs_utils.py
@@ -15,6 +15,7 @@
 """Utils for interacting with GCS paths."""
 
 import functools
+import os
 from urllib import parse
 from etils import epath
 
@@ -26,7 +27,19 @@ def is_gcs_path(path: epath.Path) -> bool:
   return path.as_posix().startswith(_GCS_PATH_PREFIX)
 
 
-def parse_gcs_path(path: epath.PathLike) -> tuple[str, str]:
+def parse_gcs_path(
+    path: epath.PathLike, add_trailing_slash: bool = True
+) -> tuple[str, str]:
+  """Parses a GCS path into bucket and path within the bucket.
+
+  Args:
+    path: The GCS path to parse (e.g., "gs://my-bucket/path/to/object").
+    add_trailing_slash: Whether to ensure the returned path has a trailing
+      slash.
+
+  Returns:
+    A tuple containing the bucket name and the path within the bucket.
+  """
   parsed = parse.urlparse(str(path))
   assert parsed.scheme == 'gs', f'Unsupported scheme for GCS: {parsed.scheme}'
   # Strip the leading slash from the path.
@@ -34,9 +47,33 @@ def parse_gcs_path(path: epath.PathLike) -> tuple[str, str]:
   if standardized_path.startswith('/'):
     standardized_path = standardized_path[1:]
   # Add a trailing slash if it's missing.
-  if not standardized_path.endswith('/'):
+  if add_trailing_slash and not standardized_path.endswith('/'):
     standardized_path = standardized_path + '/'
   return parsed.netloc, standardized_path
+
+
+def get_kvstore_for_gcs(ckpt_path: str):
+  """Constructs a TensorStore kvstore spec for a GCS path.
+
+  Args:
+    ckpt_path: A GCS path of the form gs://<bucket>/<path>.
+
+  Returns:
+    A dictionary containing the TensorStore kvstore spec.
+
+  Raises:
+    ValueError: if ckpt_path is not a valid GCS path.
+  """
+  gcs_bucket, path_without_bucket = parse_gcs_path(
+      ckpt_path, add_trailing_slash=False
+  )
+  # TODO(stoelinga): Switch to gcs_grpc by default.
+  # gcs_grpc performs roughly twice as fast as gcs backend.
+  gcs_backend = os.environ.get('TENSORSTORE_GCS_BACKEND', 'gcs')
+  spec = {'driver': gcs_backend, 'bucket': gcs_bucket}
+  if path_without_bucket:
+    spec['path'] = path_without_bucket
+  return spec
 
 
 @functools.lru_cache(maxsize=32)