+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.26] - 2025-10-22
+
 ### Added
 
 - Added an option `use_compression` to `PyTreeCheckpointHandler` to